Crate xcm_primitives

The XCM primitive trait implementations

Structs

AbsoluteAndRelativeReserve

This struct offers uses RelativeReserveProvider to output relative views of multilocations However, additionally accepts a Location that aims at representing the chain part (parent: 1, Parachain(paraId)) of the absolute representation of our chain. If a token reserve matches against this absolute view, we return Some(Location::here()) This helps users by preventing errors when they try to transfer a token through xtokens to our chain (either inserting the relative or the absolute value).

AccountIdToLocation

Instructs how to convert a 20 byte accountId into a Location

AsAssetType

Converter struct implementing AssetIdConversion converting a numeric asset ID (must be TryFrom/TryInto<u128>) into a Location Value and vice versa through an intermediate generic type AssetType. The trait bounds enforce is that the AssetTypeGetter trait is also implemented for AssetIdInfoGetter

EthereumXcmTransactionV1

EthereumXcmTransactionV2

EthereumXcmTransactionV3

IsBridgedConcreteAssetFrom

Matches foreign assets from a given origin. Foreign assets are assets bridged from other consensus systems. i.e parents > 1.

ManualEthereumXcmFee

Manually sets a gas fee.

MultiNativeAsset

A ContainsPair implementation. Filters multi native assets whose reserve is same with origin.

RelativeReserveProvider

SignedToAccountId20

Enums

AvailableStakeCalls

EthereumXcmFee

Xcm transact’s Ethereum transaction configurable fee.

EthereumXcmTransaction

Xcm transact’s Ethereum transaction.

HrmpAvailableCalls

UtilityAvailableCalls

Constants

DEFAULT_PROOF_SIZE

MAX_ASSETS

MAX_ETHEREUM_XCM_INPUT_SIZE

Max. allowed size of 65_536 bytes.

Traits

AccountIdToCurrencyId

This trait ensure we can convert AccountIds to CurrencyIds We will require Runtime to have this trait implemented

AssetTypeGetter

Defines the trait to obtain a generic AssetType from a generic AssetId and vice versa

EnsureProxy

Ensure that a proxy between delegator and delegatee exists in order to deny or grant permission to do xcm-transact to transact_through_proxy.

FilterMaxAssetFee

Filters max fee for a given multiasset. It takes self (a multiasset) and runs contains in the argument multiasset Can be amalgamated into tuples. If any item returns true, it short-circuits, else false is returned.

HrmpEncodeCall

Reserve

StakeEncodeCall

UtilityEncodeCall

XcmToEthereum

XcmTransact

Functions

rs_id

Value for r and s for the invalid signature included in Xcm transact’s Ethereum transaction.

split_location_into_chain_part_and_beneficiary