Trait xcm_primitives::EnsureProxy

pub trait EnsureProxy<AccountId> {
    // Required method
    fn ensure_ok(
        delegator: AccountId,
        delegatee: AccountId,
    ) -> Result<(), &'static str>;
}

Ensure that a proxy between delegator and delegatee exists in order to deny or grant permission to do xcm-transact to transact_through_proxy.

Required Methods

fn ensure_ok( delegator: AccountId, delegatee: AccountId, ) -> Result<(), &'static str>

Object Safety

This trait is not object safe.